Spring MVC 4

+++ src/main/java/com/rom4/departments/service/impl/DepartmentServiceImpl.java
@@ -34,9 +34,14 @@ public class DepartmentServiceImpl implements DepartmentService {
     @Qualifier("departmentValidation")
     org.springframework.validation.Validator springValidator ;
 
-    public List<ObjectError> validate(Department department) {
+    @Autowired
+    @Qualifier("departmentDeleteValidation")
+    org.springframework.validation.Validator springDeleteValidator ;
+
+    public List<ObjectError> validate(Department department, String method) {
         BeanPropertyBindingResult result = new BeanPropertyBindingResult(department, "department");
-        ValidationUtils.invokeValidator(springValidator, department, result);
+        Validator val = method.equals("delete") ? springDeleteValidator : springValidator;
+        ValidationUtils.invokeValidator(val, department, result);
         List<ObjectError> errors = result.getAllErrors();
 
         if(errors.size()>0) {
@@ -59,27 +64,31 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     @Transactional
-    public void create(Department dep)  {
-       /* List<ObjectError> errors = validate(dep);
+    public void create(Department dep) throws ValidateException  {
+        List<ObjectError> errors = validate(dep, "create");
         if ( errors != null) {
-            throw new ValidateException("Employee validation error!", errors);
-        }*/
+            throw new ValidateException("Department validation error!", errors);
+        }
          dao.createDepartment(dep);
     }
 
     @Override
     @Transactional
-    public void update(Department dep) {
-        /*List<ObjectError> errors = validate(dep);
+    public void update(Department dep) throws ValidateException {
+        List<ObjectError> errors = validate(dep, "create");
         if ( errors != null) {
-            throw new ValidateException("Employee validation error!", errors);
-        }*/
+            throw new ValidateException("Department validation error!", errors);
+        }
         dao.udpateDepartment(dep);
     }
 
     @Override
     @Transactional
-    public void delete(Department department) {
+    public void delete(Department department) throws ValidateException {
+        List<ObjectError> errors = validate(department, "delete");
+        if ( errors != null) {
+            throw new ValidateException("Department validation error!", errors);
+        }
         dao.deleteDepartment(department);
     }
